requirements.txt
Pillow==9.2.0
pandas==1.5.0
numpy==1.23.3